+++ crates/ykdf-core/src/lib.rs
@@ -1,3 +1,17 @@
 mod error;
 
+pub mod context;
+pub mod derive;
+pub mod expand;
+pub mod extract;
+pub mod pipeline;
+pub mod profile;
+pub mod types;
+
+pub use self::context::Context;
+pub use self::derive::derive;
 pub use self::error::{Error, Result};
+pub use self::extract::extract;
+pub use self::pipeline::Pipeline;
+pub use self::profile::{Profile, ProfileOutput};
+pub use self::types::{ExpandedBytes, Ikm, MasterKey};